requirements.txt
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.14.0
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Migrate==2.5.2
pycparser==2.18
six==1.11.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.1
autopep8==1.4.4
pytest==4.4.1
coverage==4.5.3
Click==7.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.2
pycryptodome==3.9.4
ecdsa==0.15
python-dotenv==0.12.0
requests==2.23.0
Flask-Limiter==1.2.1
limits==1.5.1
redis==3.4.1
